pyproject.toml
[tool.poetry]
name = "nicolas-automation-tool"
version = "0.1.0"
description = ""
authors = ["Etienne Lévêque <eti.lvque@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"pdfminer.six" = "^20191110"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.761"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
